move inference code into search_inference
